+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 - **Change:** On XF 2.2.7+ when minifying JS files go with the stock minifier service (#109)
 - **Fix:** Exception thrown when the add-on assigned to permission interface has been removed (#108)
+- **Fix:** Building readme with PHP 8.1 fails (#111)
 
 ## 1.3.6 (`1030670`)